Makefile

+++ libft/ft_printf/put_conversion/put_float.c
@@ -6,7 +6,7 @@ void	put_float(t_conversion *conv, va_list valist)
 
 	if (conv->precision < 0)
 		conv->precision = 6;
-	str = handle_float(va_arg(valist, float), conv->precision);
+	str = handle_float(va_arg(valist, double), conv->precision);
 	str = handle_minwidth(conv, str, 0);
 	putstr(str);
 	free(str);
